Move -nfh group mods into event/free_heals.py
-nfh started as an off-branch of -ruin, so its group-wide sweepers were
living in event/ruination.py. They are orthogonal now (the two flags
run independently and -ruin just includes -nfh via ruin_preprocessor),
so extract the shared code into its own module.

Moved verbatim from event/ruination.py into event/free_heals.py:
- INN_COST_MULTIPLIER
- modify_inn_costs
- FREE_BED_AMBUSH_PACK / FREE_BED_DIALOG_ID / VANILLA_BED_HEAL_ADDRESS /
  RUINATION_BED_HEAL_ADDRESS / FREE_BED_LOCATIONS
- modify_free_bed_heals
- SpringEffect / SPRING_LOCATIONS / SPRING_FLASH_COLORS /
  SPRING_DIALOG_BASE
- modify_recovery_springs

event/events.py imports the three modify_* entry points from
event/free_heals.py; event/burning_house.py and event/doma_wor.py
follow for INN_COST_MULTIPLIER and the bed-ambush constants.
Ruination-only sweepers (disable_chocobo_stables,
fix_ferry_connections) stay in event/ruination.py.

Per-event -nfh code already lives in its per-event files and is
untouched. No behavioral changes.

ARCHIVE.md and event/TODO.md are repointed to the new module.
