+++ ARCHIVE.md
@@ -969,10 +969,11 @@ under `RUIN_DEFAULT_FLAGS['other']`. Users can opt out with `-no nfh`.
 - **Flag definition:** `args/misc.py` — `misc.add_argument("-nfh", "--no-free-heals", action="store_true", ...)`. Also exposed via `flags()` and `options()` ("No Free Heals").
 - **Default in `-ruin`:** `args/ruin_preprocessor.py` `RUIN_DEFAULT_FLAGS['other']` includes `'-nfh'`.
 - **Top-level call site:** `event/events.py` `mod()` calls `self.no_free_heals_mod()` immediately after `create_party_interaction_scripts` whenever `args.no_free_heals` is true. This works in both ruination and standalone contexts.
-- **Top-level method:** `event/events.py` `no_free_heals_mod()` invokes the three "global" sweepers from `event/ruination.py`:
+- **Top-level method:** `event/events.py` `no_free_heals_mod()` invokes the three "global" sweepers from `event/free_heals.py`:
   - `modify_inn_costs(maps, rom, dialogs, args)` — multiplies all paid inn costs by `INN_COST_MULTIPLIER` (`= 3`) and converts free inns at Returners Hideout and Figaro Castle into paid ones.
   - `modify_free_bed_heals(maps, dialogs, args)` — turns the six existing free bed heals into HP-only heals with a 3/8 chance of an ambush from `FREE_BED_AMBUSH_PACK = 416`.
   - `modify_recovery_springs(maps, rom, dialogs, args)` — randomises the spring effects in `SPRING_LOCATIONS` (Phantom Forest, Cave to South Figaro).
+  - `event/ruination.py` retains only ruination-specific sweepers (`disable_chocobo_stables`, `fix_ferry_connections`), which are called from `ruination_mod`, not `no_free_heals_mod`.
 
 ### Per-Event Heal Modifications (gated locally)
 
@@ -986,7 +987,7 @@ Each event file checks `args.no_free_heals` directly so the changes apply with o
 | Phantom Train restaurant | `event/phantom_train.py` `restaurant_mod()` → `ruination_restaurant_mod()` | Replaces free meal with three priced choices (Cheap 10 GP / Filling 500 GP / Chef's Special 2000 GP) with random / risky / full-heal effects respectively. |
 | Narshe school heal pot | `event/narshe_wob.py` `ruination_mod()` (also requires `ruination_mode`) | Limits the bucket to `NUM_HEALS = 3` uses, tracked via the repurposed `NARSHE_CHECKPOINT` event word. The init writes the counter only when **both** `ruination_mode` and `no_free_heals` are set, since the event word is otherwise reserved by vanilla logic. |
 | Thamasa inn pricing | `event/burning_house.py` `ruination_inn_mod()` | When `no_free_heals` is on, "strangers" path receives `1500 * INN_COST_MULTIPLIER` GP. The companion burning-house re-trigger patch (use `character_recruited(STRAGO)` instead of `MET_STRAGO_RELM`) stays unconditional inside `ruination_inn_mod` since it is part of ruination's hub architecture. |
-| Returners / Figaro free inns | `modify_inn_costs()` (`event/ruination.py`) | Replaces free sleep events with `RemoveGP` followed by branch into the original sleep code; included via `no_free_heals_mod`. |
+| Returners / Figaro free inns | `modify_inn_costs()` (`event/free_heals.py`) | Replaces free sleep events with `RemoveGP` followed by branch into the original sleep code; included via `no_free_heals_mod`. |
 
 ### Notes / Caveats
