+++ event/events.py
@@ -4,6 +4,7 @@ import instruction.field as field
 from data.map_exit_extra import exit_data, door_to_eventname
 from data.warps import Warps, WarpPoints
 from event.ruination import *
+from event.free_heals import modify_inn_costs, modify_free_bed_heals, modify_recovery_springs
 from log.verbose import vprint
 
 class Events():
